+++ karma.conf.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const webpack = require('webpack');
-const WATCH = process.argv.indexOf('--watch') > -1;
 
 module.exports = function(config) {
   config.set({
@@ -16,10 +15,6 @@ module.exports = function(config) {
     // list of files / patterns to load in the browser
     files: [
       'test/entry.ts'
-    ],
-
-    // list of files to exclude
-    exclude: [
     ],
 
     // preprocess matching files before serving them to the browser
@@ -51,7 +46,7 @@ module.exports = function(config) {
         }]
       },
       tslint: {
-        emitErrors: !WATCH,
+        emitErrors: config.singleRun,
         failOnHint: false
       },
       plugins: [
@@ -59,20 +54,21 @@ module.exports = function(config) {
           filename: null,
           test: /\.(ts|js)($|\?)/i
         })
-      ].concat(!WATCH ? [new webpack.NoErrorsPlugin()] : [])
+      ].concat(config.singleRun ? [new webpack.NoErrorsPlugin()] : [])
     },
 
     remapIstanbulReporter: {
       reports: {
         html: 'coverage/html',
-        'text-summary': null
+        'text-summary': null,
+        lcovonly: 'coverage/lcov.info'
       }
     },
 
     // test results reporter to use
     // possible values: 'dots', 'progress'
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-    reporters: ['progress', 'coverage', 'karma-remap-istanbul'],
+    reporters: ['progress', 'karma-remap-istanbul'],
 
     // web server port
     port: 9876,
@@ -84,15 +80,8 @@ module.exports = function(config) {
     // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
     logLevel: config.LOG_INFO,
 
-    // enable / disable watching file and executing tests whenever any file changes
-    autoWatch: WATCH,
-
     // start these browsers
     // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-    browsers: ['PhantomJS'],
-
-    // Continuous Integration mode
-    // if true, Karma captures browsers, runs the tests and exits
-    singleRun: !WATCH
+    browsers: ['PhantomJS']
   });
 };